requirements.txt
flask==3.0.3
langchain==0.3.1
langgraph==0.2.24
langchain-community==0.3.1
langchain-ollama==0.2.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1